requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.13.2
django-imagekit
drf-yasg
djangorestframework-simplejwt
djangorestframework
pytils
dj-rest-auth